+++ packages/sit-onyx/src/components/OnyxHeadline/OnyxHeadline.vue
@@ -1,6 +1,10 @@
 <script lang="ts" setup>
+import { computed } from "vue";
 import { SKELETON_INJECTED_SYMBOL, useSkeletonContext } from "../../composables/useSkeletonState";
+import { injectI18n } from "../../i18n";
+import { normalizeUrlHash } from "../../utils/strings";
 import OnyxSkeleton from "../OnyxSkeleton/OnyxSkeleton.vue";
+import OnyxVisuallyHidden from "../OnyxVisuallyHidden/OnyxVisuallyHidden.vue";
 import type { OnyxHeadlineProps } from "./types";
 
 const props = withDefaults(defineProps<OnyxHeadlineProps>(), {
@@ -14,9 +18,12 @@ defineSlots<{
   default(): unknown;
 }>();
 
+const { t } = injectI18n();
 const skeleton = useSkeletonContext(props);
 
-const copyHash = async (hash: string) => {
+const normalizedHash = computed(() => (props.hash ? normalizeUrlHash(props.hash) : undefined));
+
+const copyLink = async (hash: string) => {
   const { origin, pathname, search } = window.location;
   const fullUrl = `${origin}${pathname}${search}#${hash}`;
   await navigator.clipboard.writeText(fullUrl);
@@ -32,16 +39,18 @@ const copyHash = async (hash: string) => {
   <component
     :is="props.is"
     v-else
-    :id="props.hash"
+    :id="normalizedHash"
     :class="['onyx-component', 'onyx-headline', `onyx-headline--${props.is}`]"
   >
     <a
-      v-if="props.hash"
-      :href="`#${props.hash}`"
+      v-if="normalizedHash"
+      :href="`#${normalizedHash}`"
       target="_self"
       class="onyx-headline__hash"
-      @click="copyHash(props.hash)"
+      :title="t('headline.copyLink')"
+      @click="copyLink(normalizedHash)"
     >
+      <OnyxVisuallyHidden>{{ t("headline.copyLinkTo") }}</OnyxVisuallyHidden>
       <slot />
     </a>
 
@@ -86,7 +95,6 @@ const copyHash = async (hash: string) => {
 
         &::before {
           $width: 1.5rem;
-          content: "#";
           position: absolute;
           width: $width;
           left: -$width;
@@ -96,6 +104,12 @@ const copyHash = async (hash: string) => {
           background-color: inherit;
           line-height: inherit;
           border-radius: var(--border-radius) 0 0 var(--border-radius);
+
+          // the / "" is used to ignore the content for screen readers, see:
+          // https://developer.mozilla.org/en-US/docs/Web/CSS/content#alternative_text_string_counter
+          // we still set 'content: "#"' in case the browser does not support the alternative syntax
+          content: "#";
+          content: "#" / "";
         }
       }