+++ packages/vitepress-theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sit-onyx/vitepress-theme
 
+## 1.0.0-beta.117
+
+### Patch Changes
+
+- Updated dependencies [c17ccf7]
+  - sit-onyx@1.0.0-beta.116
+
+## 1.0.0-beta.116
+
+### Patch Changes
+
+- Updated dependencies [4464ff3]
+  - sit-onyx@1.0.0-beta.115
+
 ## 1.0.0-beta.115
 
 ### Patch Changes